审批，编辑等等功能

+++ src/main/java/com/chamc/process/controller/FormController.java
@@ -3,11 +3,12 @@ package com.chamc.process.controller;
 import com.chamc.process.entity.ArchiveType;
 import com.chamc.process.entity.Attachment;
 import com.chamc.process.entity.Register;
-import com.chamc.process.entity.User;
 import com.chamc.process.service.ArchiveService;
 import com.chamc.process.service.FileService;
 import com.chamc.process.service.RegistService;
+import com.chamc.process.utils.interceptor.ErrorCode;
 import com.chamc.process.utils.interceptor.NoWrapper;
+import com.chamc.process.utils.interceptor.ProcessException;
 import org.apache.tomcat.util.http.fileupload.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -42,28 +43,59 @@ public class FormController {
     @Autowired
     ArchiveService archiveService;
 
+    /**
+     * 获取档案数据
+     * @return
+     */
     @GetMapping("archives")
     public List<ArchiveType> getArchives(){
         return this.archiveService.getArchives();
     }
 
+    /**
+     * 提交申请
+     * @param register
+     * @return
+     */
     @PostMapping("form")
     public Boolean regist(Register register){
         return this.registService.regist(register);
     }
 
-
-    @GetMapping("attachments")
-    public List<Attachment> getAttachments() {
-        User user = (User) this.request.getSession().getAttribute("user");
-        return this.fileService.getAllAttachs(user.getId());
+    /**
+     * 获取用户已存在的申请
+     * @param userId
+     * @return
+     */
+    @GetMapping("editer")
+    public Register getUserRegister(Long userId){
+        return this.registService.getByUserId(userId);
     }
 
+    /**
+     * 获取附件列表
+     * @return
+     */
+    @GetMapping("attachments")
+    public List<Attachment> getAttachments(Long uid) {
+        return this.fileService.getAllAttachs(uid);
+    }
+
+    /**
+     * 上传附件
+     * @param file
+     * @return
+     */
     @PostMapping("upload")
     public List<Attachment> upload(MultipartFile file) {
         return this.fileService.upload(file);
     }
 
+    /**
+     * 下载附件
+     * @param id
+     * @return
+     */
     @GetMapping("download")
     @NoWrapper
     public ResponseEntity<byte[]> download(Long id){
@@ -100,8 +132,16 @@ public class FormController {
         }
     }
 
-    @DeleteMapping("delete")
+    /**
+     * 删除附件
+     * @param ids
+     * @return
+     */
+    @PostMapping("delete")
     public Boolean delete(Long[] ids){
+        if(ids == null || ids.length == 0){
+            throw new ProcessException(ErrorCode.IDS_CAN_NOT_BE_NUL);
+        }
         return this.fileService.deleteById(ids);
     }
